+++ Telemed/Services/IPatientService.cs
@@ -5,8 +5,8 @@ namespace Telemed.Services.Interfaces;
 public interface IPatientService
 {
     Task<List<PatientDto>> GetAllAsync();
-    Task<PatientDto?> GetByIdAsync(int id);
+    Task<PatientDto?> GetByIdAsync(long id);
     Task<PatientDto> CreateAsync(CreatePatientDto dto);
-    Task<PatientDto?> UpdateAsync(int id, UpdatePatientDto dto);
-    Task<bool> DeleteAsync(int id);
+    Task<PatientDto?> UpdateAsync(long id, UpdatePatientDto dto);
+    Task<bool> DeleteAsync(long id);
 }
