+++ Telemed/Services/IProviderService.cs
@@ -5,8 +5,8 @@ namespace Telemed.Services.Interfaces;
 public interface IProviderService
 {
     Task<List<ProviderDto>> GetAllAsync();
-    Task<ProviderDto?> GetByIdAsync(int id);
+    Task<ProviderDto?> GetByIdAsync(long id);         // long
     Task<ProviderDto> CreateAsync(CreateProviderDto dto);
-    Task<ProviderDto?> UpdateAsync(int id, UpdateProviderDto dto);
-    Task<bool> DeleteAsync(int id);
+    Task<ProviderDto?> UpdateAsync(long id, UpdateProviderDto dto); // long
+    Task<bool> DeleteAsync(long id);                 // long
 }
